+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/.idea/psak.iml" filepath="$PROJECT_DIR$/.idea/psak.iml" />
+      <module fileurl="file://$PROJECT_DIR$/../temp-psak/.idea/psak-project.iml" filepath="$PROJECT_DIR$/../temp-psak/.idea/psak-project.iml" />
     </modules>
   </component>
 </project>
