+++ mitm_core/mitm_shark.py
@@ -1,0 +1,56 @@
+#!/usr/bin/env python3
+"""
+ Copyright (c) 2017, Syslog777
+
+ All rights reserved.
+
+ Redistribution and use in source and binary forms, with or without modification,
+ are permitted provided that the following conditions are met:
+
+     * Redistributions of source code must retain the above copyright notice,
+       this list of conditions and the following disclaimer.
+     * Redistributions in binary form must reproduce the above copyright notice,
+       this list of conditions and the following disclaimer in the documentation
+       and/or other materials provided with the distribution.
+     * Neither the name of Desktop nor the names of its contributors
+       may be used to endorse or promote products derived from this software
+       without specific prior written permission.
+
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+ CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+"""
+from scapy.all import *
+
+
+class MitmShark:
+    def __init__(self, filter):
+        self.filter = filter
+
+    def arp_display(self, pkt):
+        if pkt[ARP].op == 1:  # who-has (request)
+            return 'Request: {} is asking about {}'.format(pkt[ARP].psrc, pkt[ARP].pdst)
+        if pkt[ARP].op == 2:  # is-at (response)
+            return '*Response: {} has address {}'.format(pkt[ARP].hwsrc, pkt[ARP].psrc)
+
+    def mac_summary(self):
+        return sniff(prn=lambda pkt: 'Packet from {} going to --> {}'.format(pkt.src, pkt.dst))
+
+    def arp_summary(self):
+        sniff(prn=lambda pkt: self.arp_display(pkt), filter=self.filter, store=0)
+
+    def main(self):
+        self.arp_summary()
+
+
+if __name__ == '__main__':
+    shark = MitmShark("arp")
+    shark.main()
