--- psak/mitm/mitm_input.py
@@ -1,42 +1,0 @@
-#!/usr/bin/env python3
-
-import sys
-
-
-# TODO: add pre-execution arg options
-def interrupt():
-    print("\n[*] User requested shutdown")
-    print("Exiting....")
-    sys.exit(1)
-
-
-def get_interface():
-    try:
-        interface = input("[*] Enter desired interface: ")
-        return interface
-    except KeyboardInterrupt:
-        interrupt()
-
-
-def get_victimIP():
-    try:
-        victimIP = input("[*] Enter victim IP: ")
-        return victimIP
-    except KeyboardInterrupt:
-        interrupt()
-
-
-def get_gateIP():
-    try:
-        gateIP = input("[*] Enter router IP: ")
-        return gateIP
-    except KeyboardInterrupt:
-        interrupt()
-
-
-def get_gate_port_number():
-    try:
-        portNum = input("[*] Enter gate port number: ")
-        return portNum
-    except KeyboardInterrupt:
-        interrupt()
